v. 200+22s7l

+++ src/main/scala/al/strategies/DensityWeighted.scala
@@ -28,7 +28,7 @@ case class DensityWeighted(learner: Learner, pool: Seq[Pattern], beta: Double = 
    extends StrategyWithLearnerAndMaps with MarginMeasure {
    override val toString = "Density Weighted b" + beta + " (" + distance_name + ")"
    val abr = "DW" + distance_name.take(3)
-   val id = if (distance_name == "eucl" && (beta == 1 || beta == 0.5)) 5 + (beta - 1) * 100 else throw new Error("Parametros inesperados para DW.")
+   val id = if (distance_name == "eucl" && (beta == 1 || beta == 0.5)) (5 + (beta - 1) * 100).toInt else throw new Error("Parametros inesperados para DW.")
 
 
    protected def next(mapU: => Map[Pattern, Double], mapL: => Map[Pattern, Double], current_model: Model, unlabeled: Seq[Pattern], labeled: Seq[Pattern]) = {
